Fix macOS build
'fsevents' breaks macOS build, so it must not be installed.
As it is an optional dependency, updated Makefile passes '--ignore-optional'
argument to yarn to skip it. As that also skips other optional
dependencies, those were included back into 'package.json' as regular
dependencies.
As we only need a dmg, build options now reflect that explicitly. Also
we don't want to sign the code even if we have a valid identity in the
keychain, so the 'identity' setting was set to 'null'.